chore(monitor): permit umbrella override api group
A new api_group variable is added in the global values.yaml section so
that the cluster admin or Hephy admin user can override the API group
based on what they need.

Some older clusters will not support apps/v1, we do not go out of our
way to detect them and encourage Hephy users to upgrade to at least
v1.12 when the apps/v1 group was added, but this accommodation does not
cost anything and should make the transition smoother without forcing.